Add logic for error suppression and passing configs in the standalone driver (#18)
This PR adds error suppression logic (since `singlechecker` does not
support it) and passing configs (by lifting flags from `config.Analyzer`
for better UX) to the standalone checker.

Now it is possible to specify flags for NilAway via

```
nilaway -include-pkgs=go.uber.org -include-errors-in-files=./ ./...
```
such that only packages starting with `go.uber.org` is _analyzed_, and
only errors reported in the files within the current directory is
_reported_.

Please note the differences between _analyzed_ and _reported_: even if a
package is excluded from analysis, errors might still be reported on
them if the nilness flows happens within the analyzed package, but it
includes using a struct from the excluded package. For example, if the
analyzed package constructs a struct from the excluded package, assigns
a nil to a field, and then immediately dereferences the field. In this
case, NilAway will report an error for this flow, but the location will
be on the field (that resides in the excluded package).

By default, the standalone checker sets `-include-errors-in-files` to
the current working directory, so NilAway won't report any errors on
other packages. For other linter drivers, uses will have to set this
through the drivers instead.

The
[wiki/Configuration](https://github.com/uber-go/nilaway/wiki/Configuration)
is also updated to reflect this change, and a link to the configuration
page is added to the top-level README.

This PR also adds another linting task to run NilAway on itself.

Fixes #16.