[Follow-up of PR #24] Support function contract nonnil->nonnil for no infer mode (#36)
Follow up of PR #24 , support no infer mode for function contract
nonnil->nonnil. This is the final PR of the stack.

For no infer mode, we annotate param and result at every call site of
the contracted function. We duplicate param and return sites so they
would not conflict with each other (which is done in #24 ).
Theoretically we do not need to duplicate full triggers of contracted
functions, but in order to keep the analysis phase the same between full
and no infer mode, we still duplicate (done in #24) but we just ignore
those duplicated triggers in the inference phase when we are in no infer
mode.

Implementation:
- Parse inline annotations at call sites.
- Use the parsed annotations to check for conflicts.