Enable struct object creation to be safe without struct init analysis (#54)
This PR allows struct object creation `var s S` to be considered safe
(i.e., nonnil), avoiding reporting of false positives due to incorrect
marking of `s` as uninitialized. This support was already added in
NilAway, but was guarded by the struct init flag. This PR removes that
restriction, allowing this support to be generally available.

[Closes #53]