New error message format (#28)
This PR introduces a new, improved error message format with the aim of
improving UX.

New message format:

```
Potential nil panic detected. Observed nil flow from source to dereference point:
	-> CONSUMER_POS: explanation text from producer and consumer
	-> CONSUMER_POS: explanation text from producer and consumer
        ...
	-> CONSUMER_POS: explanation text from producer and consumer
```

`conflict.go` is revamped to allow conformance to the new error message
format. For simplicity of this PR, the grouping logic is removed. It
will be added back in a follow up PR.