Remove unused fields in primitive full trigger and unused utility functions (#63)
This PR removes an unused field `ExprRepr` and an unused utility
function for it. It was added when developing the new error message
format. However, our design has evolved since then and this field is no
longer needed. This should lead to a slight reduction of build time and
artifact size overhead.

Depends on PR #62