Fix multi-package inference algorithm which incorrectly uses token.Pos for injectivity (#50)
Our multi-package inference algorithm is broken in practice for
incremental build systems like bazel:

In order to incur minimal overhead to storage, we store a stripped down
version of annotation site (i.e., `inference.primitiveSite`) in the
inference map that compactly describes the annotation site. In this
struct, we are also storing `token.Pos` as position information of the
site for injectivity.

(`token.Pos` is simply an integer that cleverly and compactly encodes
the position information in a given file set, so that you don’t have to
store `(filename, line, column)` inside each AST node to reduce memory
consumption. Whenever you need to access this information, you can
convert by querying the file set `Fset.Position(p Pos)` . To put
differently, `token.Pos` means nothing without the `Fset` context.)

Now, let's say we have analyzed upstream package, exported
`NilableValue` in the inferred map (which includes a `token.Pos` that
only makes sense in the context of the upstream package’s `Fset` ). When
analyzing the downstream package, we will first import the
`NilableValue` annotation site (with a `token.Pos` that is only
meaningful during analysis of the upstream package) and its value in our
map, and then check the dereference `upstream.NilableValue`. However,
the `token.Pos` for the dereference point is for the downstream package
and could be completely different - we won’t find it in the map, no
conflicts => no errors.

In this PR, we replace `token.Pos` with `token.Position` (i.e., a tuple
of filename, offset, line, column) to make it package-independent.
However, this still does not fix the whole problem since we won't have
complete position information about upstream sites when analyzing
downstream packages. More specifically, the upstream file in the
`analysis.Pass.Fset` will not have complete line and column information.
Instead, the
[importer](https://cs.opensource.google/go/x/tools/+/refs/tags/v0.7.0:go/internal/gcimporter/bimport.go;l=375-385;drc=c1dd25e80b559a5b0e8e2dd7d5bd1e946aa996a0;bpv=0;bpt=0)
injects 65535 fake "lines" into the file, and the object we get for
dereference `upstream.NilableValue` will still contain completely
different position information due to this hack (in practice, we have
observed that the lines for the objects are correct, but not others).
This leads to mismatches in our inference engine: we still cannot find
nilabilities of upstream objects in the imported `InferredMap` since
their Position fields in the primitive sites are different.

Therefore, in addition to the package-independent `token.Position`, we
have also added an `ObjectPath` field to the exported site, that is an
opaque name that identifies a `types.Object` relative to its package
(see
[objectpath](https://pkg.go.dev/golang.org/x/tools/go/types/objectpath)
package for more details). This is then used to match the corresponding
upstream objects in downstream analysis, and fix the positions
information according to the imported positions from upstream analyses.

Implementation-wise, this PR contains the following changes:

* We replace `token.Pos` with `token.Position` and add an `ObjectPath`
field in primitive sites.
* We add a `primitivizer` struct that handles caching, and fixing logic
for upstream objects. Whenever we need to convert full trigger or
annotation sites to their primitive forms, we _must_ call methods on
this struct to get the correct primitive sites.
* We trim the bazel-specific prefix of `filename` in `token.Position` in
primitive sites since bazel will create a sandbox with a random prefix
for each build.

Depends on PR #63