Update x/exp to newest version (#96)
The new x/exp SortFunc returns an int instead of a bool.

This also upgraded x/tools (containing the analysis framework) to
v0.15.0. This includes various optimizations for the linter driver
framework. Most notably, it includes fact storing logic optimization
(see this
[commit](https://github.com/golang/tools/commit/d75c38746e7bc937ac6208cc0549fd192f0d4b48))
such that imported facts are not re-exported, which should bring some
performance improvements, as well as artifact size reduction.

This PR is a superset of #89.