Fix nondeterminism when filtering triggers for error returns (#186)
We have observed nondeterministic behaviors from NilAway when analyzing
the standard libraries.

When we process the full triggers for a particular package, we first
filter out the non-error triggers for the error-returning functions that
are dependent on the nilability of the error, process the other
triggers, and later removing some non-error triggers if the nilabilities
of the error are known due to the partial processing. This helps reduce
a lot of false positives.

However, we are using a set (map) to store the non-error triggers for
error-returning functions since it makes later deletion faster.
Eventually, we call `maps.Keys` to convert the map back to a slice to
continue processing the triggers.

This brings randomness, and NilAway currently depends on stable
processing orders of triggers for deterministic error reporting.

This PR updates `FilterTriggersForErrorReturn` function to return the
deleted triggers in a map instead. Then in our inference engine, we
store the triggers in slices, and eventually re-iterate the ordered
slice (and reference the deleted trigger map) to construct the final
(ordered) filtered triggers. By doing this, we should have similar
performance as the original implementation.