Remove timeout for backpropagation (#188)
We added a timeout for backpropagation (10s) for the analysis of any
particular function some time ago for keeping the build time SLAs since
we observed long analysis time for some particular structures (where the
backpropagation runs for many iterations).

@shubhamugare had thoroughly analyzed the problem and implemented
optimizations in PR #5, which drastically improved the performance of
NilAway's backpropagation algorithm. Meanwhile, the time-based bound
showed its own problem, including but certainly not limited to (1)
difficulty for debugging, (2) breaking build hermeticity, (3) flaky CI
since github CI machines are heavily resource-shared, and more.

After thorough testing in our internal repository, we see no instances
of such timeout being hit. This PR removes this timeout completely. Note
that we still keep the heuristic upper bound on backpropagation
iterations based on the number of blocks in the CFG as a final safety
net (which shouldn't be hit either). This PR also takes this opportunity
to refactor and move this check to the corresponding packages for better
code organization.