Support for handling of `ok` form for functions/methods (#157)
This pull request introduces support for the `ok` form in both
user-defined and library functions and methods. The implementation
addresses false positives, such as those identified in issue #77.

Currently, the feature is designed to handle explicit boolean returns,
specifically in the form of `return r0, r1, ..., true`. Support for
expression-based returns (e.g., `return r0, r1, ..., flag` or `return
r0, r1, ..., isOk()`) is tricky, as tracking boolean types is currently
not supported in NilAway. We can handle this scenario in the future.

[Closes #77 ]
[Depends on #156 ]