Add a helper function that wraps the `Analyzer.Run` for panics (#205)
We strive to _never_ panic in NilAway: we have deferred function calls
that recover from any panics and converts it to an error, which
eventually gets converted to a diagnostic with "INTERNAL PANIC" prefix
for users to be aware (but not to stop the build/linting process etc.).

However, we were adding the same deferred function block for _every_
sub-analyzer we have, which introduces maintainability burden.

This PR simplifies such handling by introducing a helper function that
wraps around the `Analyzer.Run` function, which:

* handles panics;
* puts the error in a field in `Result` struct to be sent via the return
value, such that any error won't stop the analysis process. The
upper-level analyzer will decide what to do with the error (for future
graceful recovery). For now, the upper-level analyzers simply early
return if there is an error, and the accumulation analyzer will convert
the errors to be diagnostics.

Tests that intentionally cause a panic inside the analyzers are also
added for each sub-analyzer to ensure that it won't crash the program
(i.e., it is properly recovered and converted to errors).

We also take this opportunity to add a few missing license headers to
some files.