Fix for loop range handling of basic types and test refactoring (#223)
[Go 1.22](https://tip.golang.org/doc/go1.22) now allows ranging over
integers. NilAway was not handling this correctly and throwing a panic.
This PR fixes that problem. I also took the opportunity to create a new
`looprange` test file that now hosts all the range related test cases
pulled in from other files and new test cases for checking the support
for basic types.

[closes #221 ]