Revisiting non-conditional short-circuit handling (#225)
This PR revisits the non-conditional short-circuit handling. Following
are the changes:

- The main change is to repurpose the `AddNilCheck` function to trigger
the handling, instead of the earlier custom (and duplicated)
implementation in `asNilCheckExpr`.
- `AddNilCheck` is strengthened to also be able to handle negations.
E.g., `!(x == nil)`
- The added benefit of using `AddNilCheck()` is also that it already has
support for `len()` checks, which the non-conditional handling can
leverage right away. A new extensive set of test cases for len checks
(as pointed in issue #92) is also added.