Fix consequent map access (with literal index) (#228)
This PR fixes the false positive reported for consequent map accesses as
shown in the example below.
```
if _, ok := mp[0]; !ok {
	mp[0] = new(int)
}
_ = *mp[0] // False positive was reported here
```

The PR extends the logic for applying guarding behavior to children
assertion nodes. In the map access case, it means extending the handling
to `indexAssertionNode` `mp[0]`, which appears as the child of
`varAssertionNode` `mp` in the assertion tree. Before this PR, NilAway
assumed the use ok-form for map access (`v, ok := mp[0]`), and hence was
only applying the guarding behavior to only the topmost node,
`varAssertion` `v` in this case.

(Note that this PR can currently only support literal indices. Support
for non-literal indices is coming soon.)