Refactor CFG preprocess logic (#249)
This PR refactors the CFG preprocess logic. Specifically, we have
extracted the rich check effect construction out of the preprocessing
logic such that it is now in its standalone file. This offers a lot
better modularity for the already-enormous `assertiontree` package.

Now in `BackpropAcrossFunc` function we are calling preprocess and
rich-check-effect-generating functions separately.