Implement fact export for manually-written and inferred function contracts (#240)
This PR implements the fact export for manually-written and inferred
function contracts. Specifically it enhances the `functioncontracts`
sub-analyzer, which imports contracts from upstream packages,
parses/infers the contracts for the local package, combine the contract
map and return it. It now also exports the contracts for the _exported_
functions in the _local_ package to be consumed by its downstream
packages.

Tests are also added to ensure the fact export and import logic works.