Fix for making rich check effect handling unbiased for return statement ordering (#254)
This PR fixes the bug which made rich check effect handling sensitive to
return statement ordering, resulting in false positives for certain
cases. The effect of ordering of statements is illustrated in the
example below.

```
// Problematic
func retPtrErr() (*int, error) {
	if cond {
		return new(int), retNil()
	}
	return nil, retErr()
}

// Works fine
func retPtrErr() (*int, error) {
	if cond {
	       return nil, retErr()	
	}	
        return new(int), retNil()
}
```

The problem is with no entry being created in `inferredMap` for non-nil
error return from `retErr()` in `StoreImplication`, since we don't
propagate non-nilness forward. This creates a problem when the anonymous
function parameter in `FilterTriggersForErrorReturn` analyzes for
nilability of return sites, where it incorrectly assumes that absence in
`inferredMap` implies unknown nilability. This PR corrects that logic.