Refine the handling of pointers in `TypeAsDeepType` (#251)
This PR refines the handling of `*types.Pointer` in `TypeAsDeepType` to
only consider a pointer type deep if its underlying elements are of a
deep type. The implication of always considering a pointer deep was that
for type like `var x *int; foo(x)`, unnecessary deep triggers were being
created. The change in this PR reduces the creation of these redundant
triggers.