Add handling for `ok` in hook.ReplaceConditional` (#311)
In our `hook.ReplaceConditional` we have support for replacing a
conditional with another since we know it has certain implications.
E.g., `errors.As(..., &targetErr)` actually means `errors.As(...,
&targetErr) && targetErr != nil`.

This PR adds support for cases where the value is stored in the init
part of the if statement, e.g., `if ok := errors.As(...); ok { ... }`.

Note that the handling is limited: we simply check the last node within
the current block. This would lead to missing cases like `if ok :=
errors.As(...); dummy && ok { ... }` since it would be canonicalized to
`if dummy { if ok { ... } }`, and we won't find the definition of `ok`
within the current block anymore (it would work if `ok` is actually
placed at the beginning, e.g., `if ok && dummy {...}`.

This is weird, but it is the simplest way to implement such support for
now (and we assume it would be a common case to place `ok` as the first
component of the expression anyways).

A more general handling to get the latest declaration with the help of
SSA analyzer will be done in follow-up PR.