Conditionally export the nolint ranges when available (#322)
In #320 we added support for cross-package nolint suppression by storing
the nolint ranges as facts.

However, we are storing the facts even if there are no `//nolint`
suppressions in the package. Functionally it is fine since having an
empty `[]Range` is OK. However, it would introduce unnecessary codec /
storage cost.

Even more so, since we are storing `Fact` for every package regardless,
it seems to have triggered a weird golangci-lint driver bug (see #321).

This PR fixes the logic to only export when there are `//nolint` ranges
to export.