Disable error grouping for unit tests (#327)
With error grouping for unit tests, some error messages could be grouped
together. This could lead to problems where all errors are grouped into
a single source:

```
var nilable *int

switch  {
case "case 1 should have an error":
case "case 2 should NOT have an error":
}
```

the error is grouped since they have the same nil source. However, if
case 2 does indeed report an error, we will not be able to detect such
discrepancy due to error grouping.

This PR disables the error grouping for unit test (and extracted a test
package `errormessage` to `grouping` since it mainly tests the error
messages w/ grouping enabled).

With this, we found a bug where we were not handling `testing.TB.Fatal`
correctly, which will be fixed in a follow-up PR.