Add hooks for no-return functions (#328)
This PR adds hooks for no-return functions. It is used to model certain
3rd party or stdlib functions where the control flow construction is not
able to infer the function is no-return. For example:

1. `zap.Fatal`-related: they have complex logic that eventually calls a
hook that is almost always configured to just panic (but we cannot infer
that purely from code).

2. `testing.TB.Fatal`-related: they are interface methods without
implementations.

When such calls are detected, we will restructure the CFG to remove all
nodes after the no-return call, and mark the successor block of the
current block to be nil.