Support for unnamed anonymous error returning functions (#332)
This PR extends the functionality of #326 by adding support for error
return functions defined as unnamed anonymous functions. Example:

```
func test() {
	x, err := func() (*int, error) {
		if dummy {
			return nil, &myErr2{}
		}
		return new(int), nil
	}()
	if err != nil {
		return
	}
	_ = *x  // safe dereference
}
```

The PR makes primarily 3 changes:

- generalize `NodeTriggersFuncErrRet` in `rich_check_effect.go` to work
on function signatures.
- update case of *ast.CallExpr in `parse_producer_expr.go` to handle
`*ast.FuncLit`. It is guarded by the `EnableAnonymousFunc` flag. Note
that in the absence of the flag, the PR employs a temporary fix to
suppress the errors, which means that we don't report false positives,
but we also don't report true positives, leading to false negatives.

---------

Co-authored-by: Yuxin Wang <yuxinw@uber.com>