Create an `*analysishelper.EnhancedPass` to host helper methods related to `*analysis.Pass` (#349)
This PR introduces a drop-in replacement for `*analysis.Pass` that
provides more helper methods that are directly attached to
`*analysis.Pass`. These helper methods usually provide logic that spans
different areas (e.g., a combination of types and ast, like `IsZero`
that takes an `ast.Expr` as input but checks if it is a compile-time
constant via `types`).

This is different than pure helper functions in `asthelper` /
`typeshelper` etc. that are only related to the particular package
(i.e., no cross-area functionality).

As part of this, we are lifting `isIntZero` helper that is directly
attached to `RootAssertionNode` to this `EnhancedPass`. Moreover, since
it is a drop-in replacement for `*analysis.Pass`, we replace most sites
within the NilAway codebase with `*analysishelper.EnhancedPass` such
that we can enjoy the new helper methods in the future without refactors
at all.

This is meant to be the initial PR that does the refactor, more helper
methods will be refactored / added to this in the future.