Check in documentations into `docs` directory (#351)
This PR checks in the existing wiki documentations into the source
repository to allow source controls (we will be using it as single
source of truth going forward).

Not only does this allow source controls (PRs can now be made against
docs), but it also enables AI agents to work more effectively when
working in NilAway repository. As part of this effort, we are checking
in `CLAUDE.md` that simply references the `docs/Developing.md` file.