Suppress generic rich-check-effect functions error reporting (#354)
This PR addresses the false positives that were reported for generic
rich-check-effect functions. Example:
```
func GenericFunc[T any]() (*T, error) {
	var resp T
	if someCond {
		return nil, errors.New("some error")
	}
	return &resp, nil
}

func CallGenericFunc() {
        v, err := GenericFunc[int]()
	if err != nil {
		return
	}
	print(*v)  // FP reported here: unassigned variable `v` dereferenced
}
```

Currently we do not handle generic rich-check-effect function calls in
NilAway. So as a temporary solution, we suppress reporting of errors
related to generic rich-check-effect functions. Note that this does help
suppress false positives, but it also means that we don't report true
positives either. We should fix this in the future when we add support
for generics in NilAway.