Allow complex length check conditions to determine the slice being nonnil (#348)
We can have different length checks in Go code i.e., `0 < len(a) - 1`
instead of simply `0 < len(a)` which would imply that `a` is nonnil.
However, NilAway currently only reasons about simple cases like `0 <
len(a)`. This PR adds the ability to reason about complex length check
conditions by traversing the AST node and extracting the `len(a)` node.

Note that we maintain an unsound assumption that the length check yields
nonnil `a` (unless explicitly checking for `len(a) == 0`) to reduce
false positives.

Fixes #346