Add support for built-in conversion functions used as index expr (#358)
This PR adds support for built-in conversion functions, such as
`string(x)`, used as index expr. Currently, without this support,
NilAway reports a false positive for the example below.
```
func test(v uint8) {
		m := make(map[string]*int)
		if m[string(v)] != nil {
			_ = *m[string(v)]  // false positive was reported here
		}
}
```

Go compiler doesn't treat `string(x)` as a call to a built-in function
named “string”, but instead it's considered to be a conversion using the
predeclared type string. Therefore, our existing logic of
`*types.BuiltIn` did not cover this case. We had to add special support
for this case by checking for
`pass.TypesInfo.ObjectOf(ident).Type().(*types.Basic)`.

[Closes #357 ]