Model `strings.Split` which always returns a non-nil slice. (#374)
`strings.Split` always returns a non-nil slice. In most cases it returns
a slice with at least length 1, with the only exception being the case
where both arguments to `strings.Split` are empty [1], in which case the
result is a _non-nil_ empty slice. So, here we make a conscious
trade-off to reduce false positive by treating the result to be non-nil
altogether (which is technically correct), where the exception should be
both rare and out of scope for NilAway.

We are also renaming one test file for better readability.

Fixes #107