Reorganize agents files (#376)
We are adding `AGENTS.md` and `CLAUDE.md` simply as symlinks to
`docs/developing.md` such that popular AI agents have initial context
when working on NilAway repository.

We are also adding a simple paragraph in `Developing` section in README
to direct potential developers to the correct doc, with additional
guidance on using AI agents.